+++ src/Components/Authority.php
@@ -49,6 +49,8 @@ final class Authority extends Component implements AuthorityInterface
     private $port;
 
     /**
+     * @deprecated since version 2.3.0 use a more appropriate named constructor.
+     *
      * New instance.
      *
      * @param mixed|null $authority
@@ -140,6 +142,54 @@ final class Authority extends Component implements AuthorityInterface
         $authority = $uri->getAuthority();
         if ('' === $authority) {
             return new self();
+        }
+
+        return new self($authority);
+    }
+
+    /**
+     * Create a new instance from null.
+     */
+    public static function createFromNull(): self
+    {
+        return new self();
+    }
+
+    /**
+     * Create a new instance from a string.
+     */
+    public static function createFromString(string $authority = ''): self
+    {
+        return new self($authority);
+    }
+
+    /**
+     * Create a new instance from a hash of parse_url parts.
+     *
+     * Create an new instance from a hash representation of the URI similar
+     * to PHP parse_url function result
+     *
+     * @param array<string,null|int|string> $components
+     */
+    public static function createFromComponents(array $components): self
+    {
+        $components += ['user' => null, 'pass' => null, 'host' => null, 'port' => null];
+
+        $authority = $components['host'];
+        if (null !== $components['port']) {
+            $authority .= ':'.$components['port'];
+        }
+
+        $userInfo = null;
+        if (null !== $components['user']) {
+            $userInfo = $components['user'];
+            if (null !== $components['pass']) {
+                $userInfo .= ':'.$components['pass'];
+            }
+        }
+
+        if (null !== $userInfo) {
+            $authority = $userInfo.'@'.$authority;
         }
 
         return new self($authority);