+++ src/Components/HierarchicalPath.php
@@ -63,6 +63,11 @@ final class HierarchicalPath extends Component implements SegmentedPathInterface
     private $segments;
 
     /**
+     * @deprecated 2.3.0
+     *
+     * @see ::createFromString
+     * @see ::createFromPath
+     *
      * New instance.
      *
      * @param mixed|string $path
@@ -80,6 +85,16 @@ final class HierarchicalPath extends Component implements SegmentedPathInterface
         }
 
         $this->segments = explode(self::SEPARATOR, $segments);
+    }
+
+    public static function createFromString(string $path = ''): self
+    {
+        return self::createFromPath(new Path($path));
+    }
+
+    public static function createFromPath(PathInterface $path): self
+    {
+        return new self($path);
     }
 
     /**