+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All Notable changes to `League\Uri\Components` will be documented in this file
 
+## Next - TBD
+
+### Added
+
+- `Authority::createFromNull`
+- `Authority::createFromString`
+- `Authority::createFromComponents`
+- `DataPath::createFromFilePath`
+- `DataPath::createFromString`
+- `Domain::createFromString`
+- `Domain::createFromHost`
+- `HierarchicalPath::createFromPath`
+- `HierarchicalPath::createFromString`
+
+### Fixed
+
+- Fix RFC1738 encoding in `QueryString::build`
+- Improve `UriModifier` using the newly added named constructors.
+
+### Deprecated
+
+- `Authority::__construct` replaced by `Authority` named constructors
+- `DataPath::createFromPath` replaced by `DataPath::createFromFilePath`
+- `Domain::__construct` replaced by `Domain` named constructors
+- `HierarchicalPath::__construct` replaced by `HierarchicalPath` named constructors
+
+### Remove
+
+- None
+
 ## 2.2.1 - 2020-02-09
 
 ### Added